// src/app/model/test.ts
export interface test {
    id2: number;
    inputDate: string;
    inputDay: string;
    inputTime: string;
    inputLocation: string;
    inputTurbidity: string;
    inputTemperature: string;
    inputColour: string;
    inputTasteAndOdour: string;
    inputSolids: string;
    inputEC: string;
    inputpH: string;
    inputChloride: string;
    inputChlorine: string;
    inputSulphate: string;
    inputNitrogen: string;
    inputFluoride: string;
    inputFe_Mn: string;
    inputCu_Zn: string;
    inputHardness: string;
    inputDO: string;
    inputBOD: string;
}